@@ -49,9 +49,9 @@
     "lodash-contrib": "^4.1200.1",
     "notistack": "^2.0.3",
     "pluralize": "^8.0.0",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-device-detect": "^2.2.2",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
     "react-draft-wysiwyg": "^1.14.7",
     "react-dropzone": "^14.2.1",
     "react-image-file-resizer": "^0.4.8",
