@@ -1,7 +1,12 @@
-import React, { useMemo } from 'react'
-import _ from 'lodash'
+import React, { useMemo } from 'react';
+import _ from 'lodash';
 
-const DOTS = '...'
+export const DOTS = '...';
+
+const range = (start, end) => {
+  let length = end - start + 1;
+  return Array.from({ length }, (_, idx) => idx + start);
+};
 
 export const usePagination = ({
   totalCount,
@@ -16,17 +21,13 @@ export const usePagination = ({
     const totalPageNumbers = siblingCount + 5;
 
     /*
-      Case 1:
       If the number of pages is less than the page numbers we want to show in our
       paginationComponent, we return the range [1..totalPageCount]
     */
     if (totalPageNumbers >= totalPageCount) {
-      return _.range(1, totalPageCount);
+      return range(1, totalPageCount);
     }
 
-    /*
-      Calculate left and right sibling index and make sure they are within range 1 and totalPageCount
-    */
     const leftSiblingIndex = Math.max(currentPage - siblingCount, 1);
     const rightSiblingIndex = Math.min(
       currentPage + siblingCount,
@@ -34,7 +35,9 @@ export const usePagination = ({
     );
 
     /*
-      We do not show dots just when there is just one page number to be inserted between the extremes of sibling and the page limits i.e 1 and totalPageCount. Hence we are using leftSiblingIndex > 2 and rightSiblingIndex < totalPageCount - 2
+      We do not want to show dots if there is only one position left 
+      after/before the left/right page count as that would lead to a change if our Pagination
+      component size which we do not want
     */
     const shouldShowLeftDots = leftSiblingIndex > 2;
     const shouldShowRightDots = rightSiblingIndex < totalPageCount - 2;
@@ -42,34 +45,24 @@ export const usePagination = ({
     const firstPageIndex = 1;
     const lastPageIndex = totalPageCount;
 
-    /*
-      Case 2: No left dots to show, but rights dots to be shown
-    */
     if (!shouldShowLeftDots && shouldShowRightDots) {
       let leftItemCount = 3 + 2 * siblingCount;
-      let leftRange = _.range(1, leftItemCount);
+      let leftRange = range(1, leftItemCount);
 
       return [...leftRange, DOTS, totalPageCount];
     }
 
-    /*
-      Case 3: No right dots to show, but left dots to be shown
-    */
     if (shouldShowLeftDots && !shouldShowRightDots) {
-
       let rightItemCount = 3 + 2 * siblingCount;
-      let rightRange = _.range(
+      let rightRange = range(
         totalPageCount - rightItemCount + 1,
         totalPageCount
       );
       return [firstPageIndex, DOTS, ...rightRange];
     }
 
-    /*
-      Case 4: Both left and right dots to be shown
-    */
     if (shouldShowLeftDots && shouldShowRightDots) {
-      let middleRange = _.range(leftSiblingIndex, rightSiblingIndex);
+      let middleRange = range(leftSiblingIndex, rightSiblingIndex);
       return [firstPageIndex, DOTS, ...middleRange, DOTS, lastPageIndex];
     }
   }, [totalCount, pageSize, siblingCount, currentPage]);
